requirements.txt
nltk
python-docx
torch
tiktoken
ollama
chromadb
openai
langchain
langchain-core
langchain-openai
langchain-community
langchain_text_splitters
unstructured
unstructured[all-docs]
unstructured[docx,pptx]
pypandoc-binary